Fix PRINTF_ARGS() for print_()
Previously the PRINT_ARGS() macro pointed to the first parameter which
actually wasn't the format string, but the context.
This bug was found after editing the definition of PRINTF_ARGS():
previously before being defined it checked for the presence of the
definition of either COMPILER_CLANG or COMPILER_GCC, instead of
checking for the presence of __GNUC__ (copy-paste mistake).